Conflicts:
	bash_aliases

+++ setup_dotfiles.sh
@@ -37,9 +37,29 @@ if [[ "${TRUE_HOST}" == "dena" ]]; then
 	declare -a files=(.bash_aliases .vimrc .tmux.conf .screenrc .pathrc .modulefiles .vncrc .gdbinit .dircolors .sqliterc .ctags)
 elif [[ "${TRUE_HOST}" != "" ]]; then
 	# We're on env can machines
-	declare -a files=(.bash_aliases .vimrc .tmux.conf .screenrc .pathrc .vncrc .gdbinit)
+	declare -a files=(.bash_aliases .pathrc .vncrc .gdbinit)
 else
-	declare -a files=(.zshrc .bashrc .bash_aliases .bash_profile .profile .login .logout .vimrc .tmux.conf .screenrc .pathrc .modulefiles .vncrc .gdbinit .dircolors .sqliterc .ctags)
+	declare -a files=(.zshrc .bashrc .bash_aliases .bash_profile .profile .login .logout .modulefiles .vncrc .gdbinit .dircolors)
+fi
+
+# Check if our environment supports these
+if [[ "$(which vim)" != "" ]]; then
+	files+=('.vimrc')
+fi
+if [[ "$(which tmux)" != "" ]]; then
+	files+=('.tmux.conf')
+	if [[ ! -e ${h}/.tmux ]]; then
+		git clone https://github.com/tmux-plugins/tpm ${h}/.tmux/plugins/tpm
+	fi
+fi
+if [[ "$(which screen)" != "" ]]; then
+	files+=('.screenrc')
+fi
+if [[ "$(which sqlite3)" != "" ]]; then
+	files+=('.sqliterc')
+fi
+if [[ "$(which ctags)" != "" ]]; then
+	files+=('.ctags')
 fi
 
 # .config/autokey
@@ -77,7 +97,12 @@ for f in ${files[@]}; do
 done;
 
 cd $h
-ln -s .vimrc .nvimrc
+if [[ -e .vimrc ]]; then
+	ln -s .vimrc .nvimrc
+fi
+if [[ -e .modulefiles ]]; then
+	ln -s .modulefiles/.modulerc ./
+fi
 
 # Can no longer to this as I'm typically using zsh
 # and this is writting in bash.  I have to keep it
